+++ tests/org.dataflowanalysis.converter.tests/src/org/dataflowanalysis/converter/tests/BehaviorTest.java
@@ -3,9 +3,7 @@ package org.dataflowanalysis.converter.tests;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 
 import org.dataflowanalysis.converter.BehaviorConverter;
